+++ api/controllers/hosts/edit.js
@@ -126,9 +126,9 @@ module.exports = {
             } else if (hostRecord.answerCalls) {
               status = 3
             }
-            await sails.models.status.changeStatus([{ name: `host-${sh.unique(hostRecord.host + sails.config.custom.hostSecret)}`, label: `Host ${hostRecord.friendlyname}`, status: status, data: `Host / DJ Controls is offline. Please ensure the application is running and connected to the internet.${additionalData}` }])
+            await sails.helpers.status.change.with({ name: `host-${sh.unique(hostRecord.host + sails.config.custom.hostSecret)}`, label: `Host ${hostRecord.friendlyname}`, status: status, data: `Host / DJ Controls is offline. Please ensure the application is running and connected to the internet.${additionalData}` })
           } else {
-            await sails.models.status.changeStatus([{ name: `host-${sh.unique(hostRecord.host + sails.config.custom.hostSecret)}`, label: `Host ${hostRecord.friendlyname}`, status: 5, data: `Host / DJ Controls is online.` }])
+            await sails.helpers.status.change.with({ name: `host-${sh.unique(hostRecord.host + sails.config.custom.hostSecret)}`, label: `Host ${hostRecord.friendlyname}`, status: 5, data: `Host / DJ Controls is online.` })
           }
         } else {
           await sails.models.status.destroy({ name: `host-${sh.unique(hostRecord.host + sails.config.custom.hostSecret)}` }).fetch()