+++ api/helpers/break/add-underwritings.js
@@ -45,7 +45,7 @@ module.exports = {
 
       // If there are no queueUnderwritings breaks, this is an error! Bail.
       if (x === 0) {
-        await sails.models.status.changeStatus([{ name: 'underwritings', label: 'Underwritings', data: `sails.models.underwritings are not airing; There are no queueUnderwritings tasks in clockwheel breaks with a quantity greater than 0. Please add a queueUnderwritings task to at least one of the clockwheel breaks in the server configuration. You can do this in DJ Controls under admin menu -> Server Configuration -> Breaks - Hourly`, status: 2 }])
+        await sails.helpers.status.change.with({ name: 'underwritings', label: 'Underwritings', data: `Underwritings are not airing; There are no queueUnderwritings tasks in clockwheel breaks with a quantity greater than 0. Please add a queueUnderwritings task to at least one of the clockwheel breaks in the server configuration. You can do this in DJ Controls under admin menu -> Server Configuration -> Breaks - Hourly`, status: 2 })
         return exits.success()
       }
 
@@ -325,18 +325,18 @@ module.exports = {
 
       // Change underwriting statuses
       if (veryBad.length > 0) {
-        await sails.models.status.changeStatus([{ name: 'underwritings', label: 'Underwritings', data: veryBad.join(` `) + ` ` + bad.join(` `), status: 2 }])
+        await sails.helpers.status.change.with({ name: 'underwritings', label: 'Underwritings', data: veryBad.join(` `) + ` ` + bad.join(` `), status: 2 })
       } else if (bad.length > 0) {
-        await sails.models.status.changeStatus([{ name: 'underwritings', label: 'Underwritings', data: bad.join(` `), status: 3 }])
+        await sails.helpers.status.change.with({ name: 'underwritings', label: 'Underwritings', data: bad.join(` `), status: 3 })
       } else {
-        await sails.models.status.changeStatus([{ name: 'underwritings', label: 'Underwritings', data: `No underwritings are significantly behind schedule.`, status: 5 }])
+        await sails.helpers.status.change.with({ name: 'underwritings', label: 'Underwritings', data: `No underwritings are significantly behind schedule.`, status: 5 })
       }
 
       sails.log.debug(`Finished.`)
 
       return exits.success()
     } catch (e) {
-      await sails.models.status.changeStatus([{ name: 'underwritings', label: 'Underwritings', data: `Internal Error with the underwritings system. Please check the npm logs.`, status: 2 }])
+      await sails.helpers.status.change.with({ name: 'underwritings', label: 'Underwritings', data: `Internal Error with the underwritings system. Please check the npm logs.`, status: 2 })
       return exits.error(e)
     }
   }