+++ CHANGELOG.md
@@ -92,6 +92,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Moved model function directors.updateDirectors to sails.helpers.directors.update. [Issue 12](https://github.com/Lovinity/wwsu/issues/12)
  - meta.changeMeta moved to sails.helpers.meta.change with input validation for all keys via meta.template. [Issue 12](https://github.com/Lovinity/wwsu/issues/12)
  - meta['A'] replaced with meta.memory.
+ - sails.models.status.changeStatus changed to sails.helpers.status.change.
 
 ## [5.1.0] - 2019-03-15
 ### Deprecated